file_chooser_native: Fix filters
Filters have to be manually added after creating the file chooser
gobject as it has no "filter" property. Following the about dialog builder, the api is changed so that it can take multiple filters.

Fixes: https://github.com/gtk-rs/gtk4-rs/issues/151
