+++ gtk4/src/lib.rs
@@ -87,6 +87,7 @@ mod entry_completion;
 mod enums;
 mod event_controller_key;
 mod file_chooser_dialog;
+mod file_chooser_native;
 mod flow_box;
 mod functions;
 mod icon_theme;
@@ -121,6 +122,7 @@ mod widget;
 pub use application::ApplicationBuilder;
 pub use border::Border;
 pub use css_location::CssLocation;
+pub use file_chooser_native::FileChooserNativeBuilder;
 pub use functions::*;
 pub use pad_action_entry::PadActionEntry;
 pub use page_range::PageRange;
